requirements.txt
bioblend
pyyaml
planemo
